requirements.txt
opentelemetry-api==0.13b0
opentelemetry-instrumentation==0.13b0
wrapt==1.12.1
